==> CHARLS/6_Physical_measures/7_Grip_Strength.Rmd <==
---
title: "Grip strength"
output:
  html_document:
    highlight: tango
    number_sections: yes
    theme: united
    toc: yes
    toc_depth: 6
    toc_float:
      collapsed: yes
      smooth_scroll: yes
---

<!-- ########################################################## --> 

<!-- # Installation, loading packages -->
<!-- **R script:** -->

```{r setup, include=FALSE}
source("M:/WPs/WP1/Maelstrom/data_process/athlos-project.github.io/setup_laia.r")
source("M:/WPs/WP1/Maelstrom/data_process/athlos-project.github.io/CHARLS/CHARLS_ids.r")
```
<!-- ########################################################## --> 



# Description of DataSchema variable

The description of harmonised variable is the following:

* Short name: `grip`
* Variable label: `Grip strength in kg`
* Variable description: `Grip strength`
* Domain: `Physical measures`
* Value type: `continuous`

* Comments: `According to a systematic review (Roberts et al, 2011), there are some inconsistencies in the use of maximum or mean grip strength as a summary measure which limit comparison of results between epidemiological studies. With multiple attempts, the maximum grip strength will be greater than the mean value. 
Even though for right handed people, there is a 10% stronger grip in the dominant hand, this is not true for left handed or ambidextrous people, whose grip strength are more or less equal in both hands (Roberts et al, 2011). 
In light of the above, the harmonised variable "grip" is proposed to be the highest score (to the nearest kg) of all hand grip measurements in each study, regardless of hand dominance. This is following a protocol suggested by Roberts et al (2011), based on the American Society of Hand Therapists (ASHT) protocol.`

* Category missings: 

**Code** | **Category Label**
----| ------------------
991 | CAPI/interviewer error
992 | Impute
993 | Disable to measure
994 | Not attempt/not done
995 | Does not answer
996 | Not applicable
997 | Refuse
998 | Do not know
999 | Missing

```{r lab, echo=FALSE}
ds_label <- "Grip strength"
```

# Data process

## CHARLS

### Wave 1 

#### Study-specific variable description

| **Name** | `qc003`| `qc004`| `qc005`| `qc006`|
|-|-|-|-|-|
| **Label** | `Left Hand-1` | `Right Hand-1` | `Left Hand-2` | `Right Hand-2` |
| **Table name** | `biomarkers_wave1` |`biomarkers_wave1` |`biomarkers_wave1` |`biomarkers_wave1` |
| **Categories** | `continuous` |`continuous` |`continuous` |`continuous` |
| **Missing values** | `993 = Respondent tried but was unable to complete the test` | `993 = Respondent tried but was unable to complete the test` | `993 = Respondent tried but was unable to complete the test`| `993 = Respondent tried but was unable to complete the test` |
| | `999 = Respondent chose not to do it` |  `999 = Respondent chose not to do it` |  `999 = Respondent chose not to do it` |  `999 = Respondent chose not to do it` |
| **Description** |  | | | |

```{r assign1, echo=FALSE}
opal.assign.table.tibble(o, 'CHARLS_gs2','CHARLS.biomarkers_wave1',
                         variables=list('qc003', 'qc004', 'qc005', 'qc006'), missings = TRUE)
CHARLS_gs2 <- opal.execute(o,'CHARLS_gs2')
opal.assign.table.tibble(o, 'CHARLS_w1_id','CHARLS.CHARLS_w1_ids',
                         variables=list('var'), missings = TRUE)
CHARLS_w1_id <- opal.execute(o,'CHARLS_w1_id')

CHARLS_gs2$oldid <- CHARLS_gs2$id
CHARLS_gs2$id <- CHARLS_ids(ids = CHARLS_gs2$oldid)

CHARLS_gs <- left_join(x = CHARLS_w1_id, y = CHARLS_gs2, by = 'id')
CHARLS_gs <- CHARLS_gs[,-2]
```

```{r local1, echo=FALSE}
vari <- CHARLS_gs$qc003
kable(Continuous_summary(var = vari, missing_values = c(993, 999))[3], caption = ds_label)
kable(Continuous_summary(vari, missing_values = c(993, 999))[2], caption = "Values")
pander(Continuous_summary(vari, missing_values = c(993, 999))$summary, caption = "Summary")
ggplot(CHARLS_gs, aes(qc003)) + geom_histogram(stat="count", fill="steelblue") + xlab(ds_label) + ylab("Frequency") + xlim(0, 500)

vari <- CHARLS_gs$qc004
kable(Continuous_summary(var = vari, missing_values = c(993, 999))[3], caption = ds_label)
kable(Continuous_summary(vari, missing_values = c(993, 999))[2], caption = "Values")
pander(Continuous_summary(vari, missing_values = c(993, 999))$summary, caption = "Summary")
ggplot(CHARLS_gs, aes(qc004)) + geom_histogram(stat="count", fill="steelblue") + xlab(ds_label) + ylab("Frequency") + xlim(0, 500)

vari <- CHARLS_gs$qc005
kable(Continuous_summary(var = vari, missing_values = c(993, 999))[3], caption = ds_label)
kable(Continuous_summary(vari, missing_values = c(993, 999))[2], caption = "Values")
pander(Continuous_summary(vari, missing_values = c(993, 999))$summary, caption = "Summary")
ggplot(CHARLS_gs, aes(qc005)) + geom_histogram(stat="count", fill="steelblue") + xlab(ds_label) + ylab("Frequency") + xlim(0, 500)

vari <- CHARLS_gs$qc006
kable(Continuous_summary(var = vari, missing_values = c(993, 999))[3], caption = ds_label)
kable(Continuous_summary(vari, missing_values = c(993, 999))[2], caption = "Values")
pander(Continuous_summary(vari, missing_values = c(993, 999))$summary, caption = "Summary")
ggplot(CHARLS_gs, aes(qc006)) + geom_histogram(stat="count", fill="steelblue") + xlab(ds_label) + ylab("Frequency") + xlim(0, 500)
```

#### Harmonisation algorithm
The harmonized variable is the higest score of all measurements and the missing values have to be recoded as follows: 

* `keep 993`
* `999 into 997`
* `NA into 999`

**R script:**

```{r harmo1}
CHARLS_gs$qc003 <- car::recode(CHARLS_gs$qc003, "993=-1; 999=-2")
CHARLS_gs$qc004 <- car::recode(CHARLS_gs$qc004, "993=-1; 999=-2")
CHARLS_gs$qc005 <- car::recode(CHARLS_gs$qc005, "993=-1; 999=-2")
CHARLS_gs$qc006 <- car::recode(CHARLS_gs$qc006, "993=-1; 999=-2")

CHARLS_ds_w1 <- tibble(id=CHARLS_gs$id)
CHARLS_ds_w1$grip <- apply(CHARLS_gs[,2:5], MARGIN = 1, FUN = function(x) max(x, na.rm = TRUE))
CHARLS_ds_w1$grip <- car::recode(CHARLS_ds_w1$grip, "-Inf=999; -1=993; -2=997")
CHARLS_ds_w1$grip <- as.integer(CHARLS_ds_w1$grip)
CHARLS_ds_w1$grip <- labelled(CHARLS_ds_w1$grip, labels = c("Missing"=999, "Do not know"=998, "Refuse"=997, "Not applicable"=996, "Does not answer"=995,  "Not attempt not done"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistics of the new harmonised variable
```{r descript1, echo=FALSE}
vari <- CHARLS_ds_w1$grip
kable(Continuous_summary(var = vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[3], caption = ds_label)
kable(Continuous_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[2], caption = "Values")
pander(Continuous_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))$summary, caption = "Summary")
ggplot(CHARLS_ds_w1, aes(grip)) + geom_histogram(stat="count", fill="steelblue") + xlab(ds_label) + ylab("Frequency") + xlim(0,100) 
```

#### Validation
There are values unusualy high: 
```{r crosstabulation1, echo=FALSE}
pander(CHARLS_ds_w1[which(CHARLS_ds_w1$grip>=100 & CHARLS_ds_w1$grip<=900),])
```

```{r save1, echo=FALSE}
grip <- CHARLS_ds_w1
save(grip, file = "../RData/w1/grip.RData")
rm(grip)
```


### Wave 2 

#### Study-specific variable description

The study-specific variable elected to be harmonised is:

| **Name** | `qc003`| `qc004`| `qc005`| `qc006`|
|-|-|-|-|-|
| **Label** | `Left Hand 1` | `Right Hand 1` | `Left Hand 2` | `Right Hand 2` |
| **Table name** | `Biomarker_wave2` |`Biomarker_wave2` |`Biomarker_wave2` |`Biomarker_wave2` |
| **Categories** | `continuous` |`continuous` |`continuous` |`continuous` |
| **Missing values** | `993 = Respondent tried but was unable to complete the test` | `993 = Respondent tried but was unable to complete the test` | `993 = Respondent tried but was unable to complete the test`| `993 = Respondent tried but was unable to complete the test` |
| | `999 = Respondent chose not to do it` |  `999 = Respondent chose not to do it` |  `999 = Respondent chose not to do it` |  `999 = Respondent chose not to do it` |
| **Description** |  | | | |



```{r assign2, echo=FALSE}
opal.assign.table.tibble(o, 'CHARLS_gs2','CHARLS.Biomarker_wave2',
                         variables=list('qc003', 'qc004', 'qc005', 'qc006'), missings = TRUE)
CHARLS_gs2 <- opal.execute(o,'CHARLS_gs2')
opal.assign.table.tibble(o, 'CHARLS_w2_id','CHARLS.CHARLS_w2_ids',
                         variables=list('var'), missings = TRUE)
CHARLS_w2_id <- opal.execute(o,'CHARLS_w2_id')

CHARLS_gs <- left_join(x = CHARLS_w2_id, y = CHARLS_gs2, by = 'id')
CHARLS_gs <- CHARLS_gs[,-2]
```

```{r local2, echo=FALSE}
vari <- CHARLS_gs$qc003
kable(Continuous_summary(var = vari, missing_values = c(993, 999))[3], caption = ds_label)
kable(Continuous_summary(vari, missing_values = c(993, 999))[2], caption = "Values")
pander(Continuous_summary(vari, missing_values = c(993, 999))$summary, caption = "Summary")
ggplot(CHARLS_gs, aes(qc003)) + geom_histogram(stat="count", fill="steelblue") + xlab(ds_label) + ylab("Frequency") + xlim(0, 500)

vari <- CHARLS_gs$qc004
kable(Continuous_summary(var = vari, missing_values = c(993, 999))[3], caption = ds_label)
kable(Continuous_summary(vari, missing_values = c(993, 999))[2], caption = "Values")
pander(Continuous_summary(vari, missing_values = c(993, 999))$summary, caption = "Summary")
ggplot(CHARLS_gs, aes(qc004)) + geom_histogram(stat="count", fill="steelblue") + xlab(ds_label) + ylab("Frequency") + xlim(0, 500)

vari <- CHARLS_gs$qc005
kable(Continuous_summary(var = vari, missing_values = c(993, 999))[3], caption = ds_label)
kable(Continuous_summary(vari, missing_values = c(993, 999))[2], caption = "Values")
pander(Continuous_summary(vari, missing_values = c(993, 999))$summary, caption = "Summary")
ggplot(CHARLS_gs, aes(qc005)) + geom_histogram(stat="count", fill="steelblue") + xlab(ds_label) + ylab("Frequency") + xlim(0, 500)

vari <- CHARLS_gs$qc006
kable(Continuous_summary(var = vari, missing_values = c(993, 999))[3], caption = ds_label)
kable(Continuous_summary(vari, missing_values = c(993, 999))[2], caption = "Values")
pander(Continuous_summary(vari, missing_values = c(993, 999))$summary, caption = "Summary")
ggplot(CHARLS_gs, aes(qc006)) + geom_histogram(stat="count", fill="steelblue") + xlab(ds_label) + ylab("Frequency") + xlim(0, 500)
```


#### Harmonisation algorithm
The harmonized variable is the higest score of all measurements and the missing values have to be recoded as follows: 

* `keep 993`
* `999 into 997`
* `NA into 999`

**R script**

```{r harmo2}
CHARLS_gs$qc003 <- car::recode(CHARLS_gs$qc003, "993=-1; 999=-2")
CHARLS_gs$qc004 <- car::recode(CHARLS_gs$qc004, "993=-1; 999=-2")
CHARLS_gs$qc005 <- car::recode(CHARLS_gs$qc005, "993=-1; 999=-2")
CHARLS_gs$qc006 <- car::recode(CHARLS_gs$qc006, "993=-1; 999=-2")

CHARLS_ds_w2 <- tibble(id=CHARLS_gs$id)
CHARLS_ds_w2$grip <- apply(CHARLS_gs[,2:5], MARGIN = 1, FUN = function(x) max(x, na.rm = TRUE))
CHARLS_ds_w2$grip <- car::recode(CHARLS_ds_w2$grip, "-Inf=999; -1=993; -2=997")
CHARLS_ds_w2$grip <- as.integer(CHARLS_ds_w2$grip)
CHARLS_ds_w2$grip <- labelled(CHARLS_ds_w2$grip, labels = c("Missing"=999, "Do not know"=998, "Refuse"=997, "Not applicable"=996, "Does not answer"=995,  "Not attempt not done"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistics of the new harmonised variable
```{r descript2, echo=FALSE}
vari <- CHARLS_ds_w2$grip
kable(Continuous_summary(var = vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[3], caption = ds_label)
kable(Continuous_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[2], caption = "Values")
pander(Continuous_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))$summary, caption = "Summary")
ggplot(CHARLS_ds_w2, aes(grip)) + geom_histogram(stat="count", fill="steelblue") + xlab(ds_label) + ylab("Frequency") + xlim(0,100) 
```

#### Validation
There are values unusualy high: 
```{r crosstabulation2, echo=FALSE}
pander(CHARLS_ds_w2[which(CHARLS_ds_w2$grip>=100 & CHARLS_ds_w2$grip<=900),])
```

```{r save2, echo=FALSE}
grip <- CHARLS_ds_w2
save(grip, file = "../RData/w2/grip.RData")
rm(grip)
```


#### Quality control
No comments.





```{r closeRsession, echo=FALSE} 
opal.logout(o)
```
