==> CHARLS/6_Physical_measures/10_sbp.Rmd <==
---
title: "Systolic blood pressure"
output: 
  html_document:
    theme: united
    highlight: tango
    toc: true
    toc_depth: 6
    toc_float:
      collapsed: true
      smooth_scroll: true
    number_sections: true
---

<!-- ########################################################## --> 

<!-- # Installation, loading packages -->
<!-- **R script:** -->



```{r setup, include=FALSE, echo=FALSE}
source("M:/WPs/WP1/Maelstrom/data_process/athlos-project.github.io/setup_laia.r")
source("M:/WPs/WP1/Maelstrom/data_process/athlos-project.github.io/CHARLS/CHARLS_ids.r")
```

<!-- ########################################################## --> 

# Description of DataSchema variable 

The description of harmonised variable is the following:
  
* Name: `sbp`
* Label: `Systolic blood pressure of participant at each wave`
* Type of variable: `countinous`
* Missings: 
    + `991 = CAPI/interviewer error`
    + `992 = Impute`
    + `993 = Disable to measure`
    + `994 = Not attempt/not done`
    + `995 = Does not answer`
    + `996 = Not applicable`
    + `997 = Refuse`
    + `998 = Do not know`
    + `999 = Missing`
* Comments: `Means of all sistolic blood pressure measures.`

```{r lab, echo=FALSE}
ds_label <- "Systolic blood pressure"
ds_label_all <- "Systolic blood pressure of participant at each wave"
```

# Data process

## CHARLS

### Wave 1 

#### Study-specific variable description

The study-specific variable elected to be harmonised is:

* Name: `qa003`
* Label: `Systolic Reading`
* Type of variable: `countinous`
* Missings
    + `993 = R tried but was unable to do it`
    + `999 = R chose not to do it`

* Name: `qa007`
* Label: `Systolic Reading`
* Type of variable: `countinous`
* Missings
    + `993 = R tried but was unable to do it`
    + `999 = R chose not to do it`

* Name: `qa011`
* Label: `Systolic Reading`
* Type of variable: `countinous`
* Missings
   + `993 = R tried but was unable to do it`
   + `999 = R chose not to do it`

```{r assign1, echo=FALSE}
opal.assign.table.tibble(o, 'CHARLS_w1_sbp2','CHARLS.biomarkers_wave1',
                         variables=list('qa003', 'qa007', 'qa011'), missings = TRUE)
CHARLS_w1_sbp2 <- opal.execute(o,'CHARLS_w1_sbp2')
opal.assign.table.tibble(o, 'CHARLS_w1_id','CHARLS.CHARLS_w1_ids',
                         variables=list('var'), missings = TRUE)
CHARLS_w1_id <- opal.execute(o,'CHARLS_w1_id')

CHARLS_w1_sbp2$oldid <- CHARLS_w1_sbp2$id
CHARLS_w1_sbp2$id <- CHARLS_ids(ids = CHARLS_w1_sbp2$oldid)

Add_indiv <- function(old, new){
  
  new_2 <- matrix(NA, nrow = dim(new)[1], ncol=dim(old)[2])
  oldm <- as.matrix(old)
  
  for(i in 1:length(new$id)){
    if(sum(new$id[i]==old$id)>=1){
      new_2[i,] <- oldm[which(new$id[i]==old$id),]
    } else {
      new_2[i,] <- c(new$id[i], NA,NA, NA, NA)
    }
  }
  
  new_22 <- data.frame(id=new_2[,1], oldid=as.numeric(new_2[,5]), qa003=as.numeric(new_2[,2]), qa007=as.numeric(new_2[,3]), qa011=as.numeric(new_2[,4]))
  new_22
}

CHARLS_w1_sbp <- Add_indiv(old = CHARLS_w1_sbp2, new = CHARLS_w1_id)

```

```{r local1, echo=FALSE}
vari <- CHARLS_w1_sbp$qa003
kable(Continuous_summary(var = vari, missing_values = c(993,999))[3], caption = ds_label)
kable(Continuous_summary(vari, missing_values = c(993,999))[2], caption = "Values")
pander(Continuous_summary(vari, missing_values = c(993,999))$summary, caption = "Summary")
ggplot(CHARLS_w1_sbp, aes(qa003)) + geom_histogram(stat="count", fill="steelblue") + xlab("Systolic Blood Pressure") + ylab("Frequency")

vari <- CHARLS_w1_sbp$qa007
kable(Continuous_summary(var = vari, missing_values = c(993,999))[3], caption = ds_label)
kable(Continuous_summary(vari, missing_values = c(993,999))[2], caption = "Values")
pander(Continuous_summary(vari, missing_values = c(993,999))$summary, caption = "Summary")
ggplot(CHARLS_w1_sbp, aes(qa007)) + geom_histogram(stat="count", fill="steelblue") + xlab("Systolic Blood Pressure") + ylab("Frequency")

vari <- CHARLS_w1_sbp$qa011
kable(Continuous_summary(var = vari, missing_values = c(993,999))[3], caption = ds_label)
kable(Continuous_summary(vari, missing_values = c(993,999))[2], caption = "Values")
pander(Continuous_summary(vari, missing_values = c(993,999))$summary, caption = "Summary")
ggplot(CHARLS_w1_sbp, aes(qa011)) + geom_histogram(stat="count", fill="steelblue") + xlab("Systolic Blood Pressure") + ylab("Frequency")
```

#### Harmonisation algorithm
The harmonized variable is the mean value between the first and second measurement of the study-specific variables, but the missing values have to be recoded as follows:

* `keep "R tried but was unable to do it" into 993`
* `999 into 997`
* `NA into 999`

It has to be taken into account that the individuals with 993 in any of the measurements it has to do the mean with the earlier measurements.

**R script**

```{r harmo1}
ids993_qa003 <- which(CHARLS_w1_sbp$qa003==993)
ids993_qa007 <- which(CHARLS_w1_sbp$qa007==993)
ids993_qa011<- which(CHARLS_w1_sbp$qa011==993)

CHARLS_w1_sbp$mean_sbp <- apply(X = CHARLS_w1_sbp[,c(3:5)], MARGIN = 1, FUN = mean)

CHARLS_w1_sbp$mean_sbp[ids993_qa003] <- 993
CHARLS_w1_sbp$mean_sbp[ids993_qa007] <- CHARLS_w1_sbp$qa003[ids993_qa007]
CHARLS_w1_sbp$mean_sbp[ids993_qa011] <- apply(X = CHARLS_w1_sbp[ids993_qa011,c(3,4)], MARGIN = 1, FUN = mean)

CHARLS_ds_w1 <- tibble(id=CHARLS_w1_sbp$id)
CHARLS_ds_w1$sbp <- car::recode(round(CHARLS_w1_sbp$mean_sbp,4), "NA=999; 999=997")
CHARLS_ds_w1$sbp <- labelled(CHARLS_ds_w1$sbp, labels = c("Missing"=999, "Do not know"=998, "Refuse"=997, "Not applicable"=996, "Does not answer"=995,  "Not attempt not done"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistics of the new harmonised variable
```{r descript1, echo=FALSE}
vari <- CHARLS_ds_w1$sbp
kable(Continuous_summary(var = vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[3], caption = ds_label)
kable(Continuous_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[2], caption = "Values")
pander(Continuous_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))$summary, caption = "Summary")
ggplot(CHARLS_ds_w1, aes(sbp)) + geom_histogram(stat="count", fill="steelblue") + xlab("Systolic Blood Pressure") + ylab("Frequency") + xlim(50,300) 
```

#### Validation
```{r crosstabulation1, echo=FALSE}
```





### Wave 2 

#### Study-specific variable description

The study-specific variable elected to be harmonised is:

* Name: `qa003`
* Label: `Systolic Reading`
* Type of variable: `countinous`
* Missings
    + `993 = R tried but was unable to do it`
    + `999 = R chose not to do it`

* Name: `qa007`
* Label: `Systolic Reading`
* Type of variable: `countinous`
* Missings
    + `993 = R tried but was unable to do it`
    + `999 = R chose not to do it`

* Name: `qa011`
* Label: `Systolic Reading`
* Type of variable: `countinous`
* Missings
   + `993 = R tried but was unable to do it`
   + `999 = R chose not to do it`

```{r assign2, echo=FALSE}
opal.assign.table.tibble(o, 'CHARLS_w2_sbp2','CHARLS.Biomarker_wave2',
                         variables=list('qa003', 'qa007', 'qa011'), missings = TRUE)
CHARLS_w2_sbp2 <- opal.execute(o,'CHARLS_w2_sbp2')
opal.assign.table.tibble(o, 'CHARLS_w2_id','CHARLS.CHARLS_w2_ids',
                         variables=list('var'), missings = TRUE)
CHARLS_w2_id <- opal.execute(o,'CHARLS_w2_id')

Add_indiv <- function(old, new){
  
  new_2 <- matrix(NA, nrow = dim(new)[1], ncol=dim(old)[2])
  oldm <- as.matrix(old)
  
  for(i in 1:length(new$id)){
    if(sum(new$id[i]==old$id)>=1){
      new_2[i,] <- oldm[which(new$id[i]==old$id),]
    } else {
      new_2[i,] <- c(new$id[i], NA, NA, NA)
    }
  }
  
  new_22 <- data.frame(id=new_2[,1], qa003=as.numeric(new_2[,2]), qa007=as.numeric(new_2[,3]), qa011=as.numeric(new_2[,4]))
  new_22
}

CHARLS_w2_sbp <- Add_indiv(old = CHARLS_w2_sbp2, new = CHARLS_w2_id)
```

```{r local2, echo=FALSE}
vari <- CHARLS_w2_sbp$qa003
kable(Continuous_summary(var = vari, missing_values = c(993,999))[3], caption = ds_label)
kable(Continuous_summary(vari, missing_values = c(993,999))[2], caption = "Values")
pander(Continuous_summary(vari, missing_values = c(993,999))$summary, caption = "Summary")
ggplot(CHARLS_w2_sbp, aes(qa003)) + geom_histogram(stat="count", fill="steelblue") + xlab("Systolic Blood Pressure") + ylab("Frequency")

vari <- CHARLS_w2_sbp$qa007
kable(Continuous_summary(var = vari, missing_values = c(993,999))[3], caption = ds_label)
kable(Continuous_summary(vari, missing_values = c(993,999))[2], caption = "Values")
pander(Continuous_summary(vari, missing_values = c(993,999))$summary, caption = "Summary")
ggplot(CHARLS_w2_sbp, aes(qa007)) + geom_histogram(stat="count", fill="steelblue") + xlab("Systolic Blood Pressure") + ylab("Frequency")

vari <- CHARLS_w2_sbp$qa011
kable(Continuous_summary(var = vari, missing_values = c(993,999))[3], caption = ds_label)
kable(Continuous_summary(vari, missing_values = c(993,999))[2], caption = "Values")
pander(Continuous_summary(vari, missing_values = c(993,999))$summary, caption = "Summary")
ggplot(CHARLS_w2_sbp, aes(qa011)) + geom_histogram(stat="count", fill="steelblue") + xlab("Systolic Blood Pressure") + ylab("Frequency")
```


#### Harmonisation algorithm
The harmonized variable is the mean value between the first, second and third measurement of the study-specific variables, but the missing values have to be recoded as follows:

* `keep "R tried but was unable to do it" into 993`
* `999 into 997`
* `NA into 999`

It has to be taken into account that the individuals with 993 in any of the measurements it has to do the mean with the earlier measurements.

**R script**

```{r harmo2}
CHARLS_w2_sbp$mean_sbp <- apply(X = CHARLS_w2_sbp[,c(2,3,4)], MARGIN = 1, FUN = mean)

CHARLS_ds_w2 <- tibble(id=CHARLS_w2_sbp$id)
CHARLS_ds_w2$sbp <- car::recode(CHARLS_w2_sbp$mean_sbp, "NA=999; 999=997")
CHARLS_ds_w2$sbp <- labelled(CHARLS_ds_w2$sbp, labels = c("Missing"=999, "Do not know"=998, "Refuse"=997, "Not applicable"=996, "Does not answer"=995,  "Not attempt not done"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistics of the new harmonised variable
```{r descript2, echo=FALSE}
vari <- CHARLS_ds_w2$sbp
kable(Continuous_summary(var = vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[3], caption = ds_label)
kable(Continuous_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[2], caption = "Values")
pander(Continuous_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))$summary, caption = "Summary")
ggplot(CHARLS_ds_w2, aes(sbp)) + geom_histogram(stat="count", fill="steelblue") + xlab("Systolic Blood Pressure") + ylab("Frequency") + xlim(50,300) 
```

#### Validation
```{r crosstabulation2, echo=FALSE}
```


# Quality estimation

Systolic bp was calculated as the mean from the 3 available variables.

```{r save, echo=FALSE}
sbp <- CHARLS_ds_w1
save(sbp, file = "../RData/w1/sbp.RData")
rm(sbp)

sbp <- CHARLS_ds_w2
save(sbp, file = "../RData/w2/sbp.RData")
rm(sbp)
```

```{r closeRsession, echo=FALSE} 
opal.logout(o)
```
